# static/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "The Cerberus is an users' data keeper. The Cerberus encrypts data and pushes it into S3.",
    "title": "Cerberus",
    "version": "1.5.2"
  },
  "basePath": "/v1",
  "paths": {
    "/accounts/{owner}/pdv-delta": {
      "get": {
        "description": "Returns PDV reward delta with reward pool",
        "produces": [
          "application/json"
        ],
        "tags": [
          "PDVRewards"
        ],
        "summary": "Get PDV reward delta of the given account",
        "operationId": "PDVRewardDelta",
        "parameters": [
          {
            "type": "string",
            "description": "account address",
            "name": "owner",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "pool",
            "schema": {
              "$ref": "#/definitions/PDVRewardDelta"
            }
          },
          "404": {
            "description": "PDV doesn't exist",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/configs/blacklist": {
      "get": {
        "description": "Returns blacklist.",
        "tags": [
          "Configs"
        ],
        "summary": "Get blacklist",
        "operationId": "GetBlacklistConfig",
        "responses": {
          "200": {
            "description": "blacklist",
            "schema": {
              "$ref": "#/definitions/Blacklist"
            }
          },
          "500": {
            "description": "internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/configs/rewards": {
      "get": {
        "description": "Returns rewards config.",
        "tags": [
          "Configs"
        ],
        "summary": "Get rewards config",
        "operationId": "GetRewardsConfig",
        "responses": {
          "200": {
            "description": "rewards config",
            "schema": {
              "$ref": "#/definitions/ObjectTypes"
            }
          },
          "500": {
            "description": "internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/images": {
      "post": {
        "security": [
          {
            "public_key": [],
            "signature": []
          }
        ],
        "description": "Resizes (1920x1080, 480x270) and saves images",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Image"
        ],
        "operationId": "Save",
        "responses": {
          "200": {
            "description": "image successfully resized and saved as HD (1920x1080) and thumbnail (480x270)",
            "schema": {
              "$ref": "#/definitions/SaveImageResponse"
            }
          },
          "500": {
            "description": "internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/pdv": {
      "post": {
        "security": [
          {
            "public_key": [],
            "signature": []
          }
        ],
        "description": "Encrypts and saves PDV",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "PDV"
        ],
        "operationId": "Save",
        "parameters": [
          {
            "description": "batch of pdv",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PDV"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "pdv was put into storage",
            "schema": {
              "$ref": "#/definitions/SavePDVResponse"
            }
          },
          "400": {
            "description": "bad request",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "401": {
            "description": "signature wasn't verified",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "403": {
            "description": "profile is banned or fraud detected",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
    "/pdv/validate": {
      "post": {
        "description": "Encrypts and saves PDV",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "PDV"
        ],
        "operationId": "Validate",
        "parameters": [
          {
            "description": "batch of pdv",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PDV"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "all pdv are valid",
            "schema": {
              "$ref": "#/definitions/ValidatePDVResponse"
            }
          },
          "400": {
            "description": "bad request",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/pdv/{owner}": {
      "get": {
        "description": "Lists PDV",
        "produces": [
          "application/json"
        ],
        "tags": [
          "PDV"
        ],
        "operationId": "List",
        "parameters": [
          {
            "type": "string",
            "example": "furya1u9slwz3sje8j94ccpwlslflg0506yc8y2ylmtz",
            "description": "PDV's address",
            "name": "owner",
            "in": "path",
            "required": true
          },
          {
            "type": "integer",
            "format": "uint64",
            "description": "id of PDV to start from",
            "name": "from",
            "in": "query"
          },
          {
            "maximum": 1000,
            "type": "integer",
            "format": "uint16",
            "description": "how many pdv will be returned",
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "List of PDV",
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "uint64"
              }
            }
          },
          "400": {
            "description": "bad request",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/pdv/{owner}/{id}": {
      "get": {
        "security": [
          {
            "public_key": [],
            "signature": []
          }
        ],
        "description": "Returns plain PDV",
        "produces": [
          "application/json"
        ],
        "tags": [
          "PDV"
        ],
        "operationId": "Get",
        "parameters": [
          {
            "type": "string",
            "description": "PDV's address",
            "name": "address",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "PDV",
            "schema": {
              "$ref": "#/definitions/PDV"
            }
          },
          "400": {
            "description": "bad request",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "401": {
            "description": "signature wasn't verified",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "403": {
            "description": "access to file is denied",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/pdv/{owner}/{id}/meta": {
      "get": {
        "description": "Returns metadata of PDV",
        "tags": [
          "PDV"
        ],
        "summary": "Get meta",
        "operationId": "GetMeta",
        "parameters": [
          {
            "type": "string",
            "description": "PDV's address",
            "name": "address",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "metadata of pdv",
            "schema": {
              "$ref": "#/definitions/PDVMeta"
            }
          },
          "400": {
            "description": "bad request",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "404": {
            "description": "PDV doesn't exist",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/profiles": {
      "get": {
        "description": "Returns profiles by addresses",
        "tags": [
          "Profile"
        ],
        "summary": "Get profiles",
        "operationId": "GetProfiles",
        "parameters": [
          {
            "example": "furya1u9slwz3sje8j94ccpwlslflg0506yc8y2ylmtz,furya1u9slwz3sje8j94ccpwlslflg0506yc8y2ylmtz",
            "description": "profile address",
            "name": "address",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "metadata of pdv",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/APIProfile"
              }
            }
          },
          "500": {
            "description": "internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "APIProfile": {
      "type": "object",
      "title": "Profile ...",
      "properties": {
        "address": {
          "type": "string",
          "x-go-name": "Address"
        },
        "avatar": {
          "type": "string",
          "x-go-name": "Avatar"
        },
        "banned": {
          "type": "boolean",
          "x-go-name": "Banned"
        },
        "bio": {
          "type": "string",
          "x-go-name": "Bio"
        },
        "birthday": {
          "type": "string",
          "x-go-name": "Birthday"
        },
        "createdAt": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "CreatedAt"
        },
        "emails": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Emails"
        },
        "firstName": {
          "type": "string",
          "x-go-name": "FirstName"
        },
        "gender": {
          "type": "string",
          "x-go-name": "Gender"
        },
        "lastName": {
          "type": "string",
          "x-go-name": "LastName"
        }
      },
      "x-go-name": "Profile",
      "x-go-package": "github.com/TessorNetwork/cerberus/internal/server"
    },
...
    "Blacklist": {
      "type": "object",
      "title": "Blacklist contains attributes of worthless pdv.",
      "properties": {
        "cookieSource": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "CookieSource"
        }
      },
      "x-go-package": "github.com/TessorNetwork/cerberus/internal/service"
    },
...
    "Fur": {
      "description": "NOTE: never use new(Fur) or else we will panic unmarshalling into the\nnil embedded big.Int",
      "type": "object",
      "x-go-package": "github.com/cosmos/cosmos-sdk/types"
    },
    "Error": {
      "type": "object",
      "title": "Error ...",
      "properties": {
        "error": {
          "type": "string",
          "x-go-name": "Error"
        }
      },
      "x-go-package": "github.com/TessorNetwork/go-api"
    },
...
    "ObjectTypes": {
      "type": "object",
      "title": "ObjectTypes contains count of each pdv type in batch.",
      "properties": {
        "advertiserId": {
          "type": "integer",
          "format": "uint16",
          "x-go-name": "AdvertiserID"
        },
        "cookie": {
          "type": "integer",
          "format": "uint16",
          "x-go-name": "Cookie"
        },
        "location": {
          "type": "integer",
          "format": "uint16",
          "x-go-name": "Location"
        },
        "profile": {
          "type": "integer",
          "format": "uint16",
          "x-go-name": "Profile"
        },
        "searchHistory": {
          "type": "integer",
          "format": "uint16",
          "x-go-name": "SearchHistoryV1"
        }
      },
      "x-go-package": "github.com/TessorNetwork/cerberus/internal/server/swagger"
    },
    "PDV": {
      "type": "object",
      "properties": {
        "version": {
          "x-go-name": "Version"
        }
      },
      "x-go-name": "PDVInterface",
      "x-go-package": "github.com/TessorNetwork/cerberus/internal/server/swagger",
      "discriminator": "version"
    },
    "PDVMeta": {
      "type": "object",
      "title": "PDVMeta contains info about PDV.",
      "properties": {
        "object_types": {
          "$ref": "#/definitions/ObjectTypes"
        },
        "reward": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "Reward"
        }
      },
      "x-go-package": "github.com/TessorNetwork/cerberus/internal/server/swagger"
    },
    "PDVRewardDelta": {
      "type": "object",
      "title": "PDVRewardDelta ...",
      "properties": {
        "delta": {
          "$ref": "#/definitions/Fur"
        },
        "pool": {
          "$ref": "#/definitions/PDVRewardsPool"
        }
      },
      "x-go-package": "github.com/TessorNetwork/cerberus/internal/server"
    },
    "PDVRewardsPool": {
      "type": "object",
      "title": "PDVRewardsPool ...",
      "properties": {
        "next_distribution_date": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "NextDistributionDate"
        },
        "size": {
          "$ref": "#/definitions/Fur"
        },
        "total_delta": {
          "$ref": "#/definitions/Fur"
        }
      },
      "x-go-package": "github.com/TessorNetwork/cerberus/internal/server"
    },
...
    "SaveImageResponse": {
      "type": "object",
      "title": "SaveImageResponse ...",
      "properties": {
        "hd": {
          "type": "string",
          "x-go-name": "HD"
        },
        "thumb": {
          "type": "string",
          "x-go-name": "Thumb"
        }
      },
      "x-go-package": "github.com/TessorNetwork/cerberus/internal/server"
    },
    "SavePDVResponse": {
      "type": "object",
      "title": "SavePDVResponse ...",
      "properties": {
        "id": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "ID"
        }
      },
      "x-go-package": "github.com/TessorNetwork/cerberus/internal/server"
    },
...
    "ValidatePDVResponse": {
      "type": "object",
      "title": "ValidatePDVResponse ...",
      "properties": {
        "invalidPDV": {
          "type": "array",
          "items": {
            "type": "integer",
            "format": "int64"
          },
          "x-go-name": "InvalidPDV"
        },
        "valid": {
          "type": "boolean",
          "x-go-name": "Valid"
        }
      },
      "x-go-package": "github.com/TessorNetwork/cerberus/internal/server"
    },
...
  },
  "securityDefinitions": {
    "public_key": {
      "description": "Blockchain account's public key",
      "type": "apiKey",
      "name": "Public-Key",
      "in": "header"
    },
    "signature": {
      "description": "Signature of request digest.\u003cbr\u003e\nDigest is sha256 sum of request: `{body as is}`+`{request uri}`.\u003cbr\u003e\nFor example:\u003cbr\u003e\nPrivate key in hex: ```cfe43c70347c7e39084612d9448f3ed86ed733a33a67de35c7e335b3c4edc37d```\u003cbr\u003e\nRequest url: ```http://localhost/v1/pdv```\u003cbr\u003e\nBody: ```{\"some\":\"file\"}```\u003cbr\u003e\nDigest will be made from ```{\"some\":\"file\"}/v1/pdv```\u003cbr\u003e\nDigest in hex:\u003cbr\u003e\n```4a1084d05820d60aee9ce600227ca2290ef63e80e5227215b58b023ec6876799```\u003cbr\u003e\nSignature in hex:\u003cbr\u003e\n```28eff4676d7839648dda925ba92d447dd7552e177a302f32681fc76278088f9f1fb98051666aa02dd80f7d9b7c01d42ea1abbb3e65de8f1fd04be7b747fb0692```\u003cbr\u003e",
      "type": "apiKey",
      "name": "Signature",
      "in": "header"
    }
  }
}